[Refactor #111]TimerManager에서 시간 설정하는 로직  TimerSettingManager로 분리
